+++ app/page.tsx
@@ -4,7 +4,7 @@ import styles from "./page.module.css";
 import Input from "./components/Input/Input";
 import Link from "next/dist/client/link";
 import Button from "./components/Button/Button";
-import { IconEnum } from "./components/Icon/Icon";
+
 import { FC } from "react";
 
 
@@ -12,6 +12,7 @@ import { FC } from "react";
 export default function Home() {
   return (
     <main className={styles.main}>
+      
       
     </main>
   );
