+++ app/styles/colors.scss
@@ -30,5 +30,6 @@ $textBlack: #050505;
 $textBlackClicked: #050505CC;
 $textSecondary: #9F9F9F;
 $textGray: #A8A8A8;
+$mainButtonHover: #FFFFFFB2;
 $textLightGray: #FFFFFFB3;
 $textWhite: #FFFFFF;
